# src/features/cartItem/cartItem.model.js
export class CartItemModel {
    constructor(productId, userId, quantity, id) {
        this.productId = productId;
        this.userId = userId;
        this.quantity = quantity;
        this._id = id;
    }



}