# src/features/user/user.model.js
export class UserModel {
    constructor(name, email, password, type, id) {
        this.name = name;
        this.email = email;
        this.password = password;
        this.type = type;
        this._id = id;
    }


    static getAll() {
        return users;
    }
}