# swaggerc.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Commerce API",
        "version": "1.0.0",
        "description": "API for managing cart items, products, orders, and users"
    },
    "servers": [
        {
            "url": "http://localhost:{port}",
            "description": "Local development server",
            "variables": {
                "port": {
                    "default": "3500",
                    "description": "The port on which the API is served"
                }
            }
        }
    ],
    "components": {
        "securitySchemes": {
            "jwt": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization",
                "description": "JWT Bearer token authentication"
            }
        },
        "schemas": {
            "CartItem": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "productID": {
                        "type": "integer"
                    },
                    "userID": {
                        "type": "integer"
                    },
                    "quantity": {
                        "type": "integer"
                    }
                }
            },
            "Product": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "desc": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "imageUrl": {
                        "type": "string"
                    },
                    "category": {
                        "type": "string"
                    },
                    "sizes": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "ratings": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "userID": {
                                    "type": "integer"
                                },
                                "rating": {
                                    "type": "number"
                                }
                            }
                        }
                    }
                }
            },
            "User": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    }
                }
            },
            "UserSignUp": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    }
                },
                "required": [
                    "name",
                    "email",
                    "password",
                    "type"
                ]
            },
            "UserSignIn": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string"
                    },
                    "password": {
                        "type": "string"
                    }
                },
                "required": [
                    "email",
                    "password"
                ]
            },
            "Order": {
                "type": "object",
                "properties": {
                    "userID": {
                        "type": "integer"
                    },
                    "totalAmount": {
                        "type": "number"
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time"
                    }
                }
            }
        }
    },
    "paths": {
        "/api/users/signup": {
            "post": {
                "summary": "Sign up a new user",
                "tags": [
                    "Users"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserSignUp"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/users/signin": {
            "post": {
                "summary": "Sign in a user",
                "tags": [
                    "Users"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserSignIn"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful sign in",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Incorrect credentials"
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "summary": "Get all products",
                "tags": [
                    "Products"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Add a new product",
                "tags": [
                    "Products"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "sizes": {
                                        "type": "string"
                                    },
                                    "imageUrl": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/products/{id}": {
            "get": {
                "summary": "Get a product by ID",
                "tags": [
                    "Products"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Product"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
        "/api/cartItems": {
            "get": {
                "summary": "Get cart items",
                "tags": [
                    "Cart Items"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CartItem"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Add item to cart",
                "tags": [
                    "Cart Items"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "productID": {
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    }
                                },
                                "required": [
                                    "productID",
                                    "quantity"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Cart is updated"
                    }
                }
            }
        },
        "/api/cartItems/{id}": {
            "delete": {
                "summary": "Delete cart item",
                "tags": [
                    "Cart Items"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cart item is removed"
                    },
                    "404": {
                        "description": "Item not found"
                    }
                }
            }
        },
        "/api/orders": {
            "post": {
                "summary": "Create an order",
                "tags": [
                    "Orders"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "summary": "Get all orders",
                "tags": [
                    "Orders"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Order"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}